Move generic launch command support into shared CLI library.
