+++ nix/.stack.nix/cardano-wallet-cli.nix
@@ -23,7 +23,10 @@
           (hsPkgs.base)
           (hsPkgs.bytestring)
           (hsPkgs.cardano-wallet-core)
+          (hsPkgs.cardano-wallet-launcher)
+          (hsPkgs.directory)
           (hsPkgs.docopt)
+          (hsPkgs.filepath)
           (hsPkgs.fmt)
           (hsPkgs.heredoc)
           (hsPkgs.http-client)
