+++ nix/.stack.nix/cardano-wallet.nix
@@ -26,10 +26,7 @@
             (hsPkgs.cardano-wallet-http-bridge)
             (hsPkgs.cardano-wallet-jormungandr)
             (hsPkgs.cardano-wallet-launcher)
-            (hsPkgs.directory)
             (hsPkgs.docopt)
-            (hsPkgs.filepath)
-            (hsPkgs.fmt)
             (hsPkgs.heredoc)
             (hsPkgs.http-client)
             (hsPkgs.iohk-monitoring)
